+++ host/src/main.tsx
@@ -1,6 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import '@party/ui/tokens.css'
+import '@party/ui/components.css'
 import './index.css'
 import App from './App.tsx'